requirements.txt
Django
django-extensions
django-webpack-loader
psycopg2
djangorestframework
markdown
django-filter
pillow
graphene-django
django-storages[google]
gunicorn
dj-database-url
django-imagekit
django-graphql-jwt
graphene-file-upload
django-debug-toolbar

